refactor(ui): extract search controls/logic into separate module
- Add src/assets/js/search.js encapsulating search UI and PDF text matching
- Remove inline search implementation from ui.js; import setupSearchControls
- Pass setPageByNumber into search module; wrap to also update URL hash
- No behavioural changes intended; build and lint pass
